lbQuery::getTableName(char* columnName) was the cause for the corrupted memory. It used a global variable to store the last result to be returned.
The problem was at creation time of the dynamic buffer. That depended on
the size of the current table name to be returned. Then the buffer was allocated
and reused, so if the reused buffer with a bigger string was used I got a buffer
overrun. This was caused by a strcpy line and using strdup prevented me from
correctly checking for an overrun.